# optimized connected component computation