--- a/bindings/c/CMakeLists.txt
+++ b/bindings/c/CMakeLists.txt
@@ -14,7 +14,7 @@ endif()
 
 target_link_libraries(
   ${PROJECT_NAME}
-  PRIVATE manifold sdf graphlite cross_section
+  PRIVATE manifold sdf cross_section
 )
 
 target_include_directories(${PROJECT_NAME} PUBLIC ${PROJECT_SOURCE_DIR}/include)